Safari-609.1.11.1/LayoutTests/fast/events/ios/autocorrect-after-image-expected.txt

The
This test verifies that inserting an auto-corrected text after an image does not remove the image.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS Observed input event of inputType insertText
PASS document.activeElement is editor
PASS editor.children.length is 1
PASS editor.children[0] instanceof HTMLImageElement is true
PASS successfullyParsed is true

TEST COMPLETE
